requirements.txt
django==3.2.3
djangorestframework==3.12.4
tensorflow>=2.4.1
psycopg2-binary>=2.8
Pillow==8.1.2
numpy==1.19.5
django-cors-headers==3.7.0